Platform And Profile version wildcard filtering (#1672)
Adding the ability to filter reports based on a platform and profile names and version numbers.
This changes the UI and the compliance-service. The compliance-service adds two new filter types with the name of platform_with_version and profile_with_version. These allow suggestions and filtering of reports with the versions combined with the name of the item.

These changes involved a migration that creates the profile full name (that is the profile title and version combined) and allows suggestions on the platform's full name. This migration works all the back to Automate 1.

Added the full name to the platform and profile fields for API requests.

Signed-off-by: Lance Finfrock <lfinfrock@chef.io>