ci: don't verify manifest in manual upgrade test
This test suite downloads the manifest from a special test scaffolding
that pretends to be packages.chef.io. Since we don't have access to
the key (and don't want access to the key) we can't sign the manifest
in this test.

Manifest verification was already disabled for deployment-service in
this test suite, this disables it for automate-cli.

Signed-off-by: Steven Danna <steve@chef.io>