cereal: introduce WorkflowName and TaskName types (#1777)
We recently shipped a bug related to passing the wrong string to the
argument of one of the cereal library's functions.  This same issue
has hit us multiple times during development. This change introduces
two types

- WorkflowName
- TaskName

and changes the public APIs of the cereal library to take these types
rather than strings.  Our belief is that this will make it hard to
make such mistakes in the future.

In total this change:

- introduces the WorkflowName and TaskName types into the cereal API
- changes all callers to work with the new API

- adds a new helper function to the SingleTaskWorkflow pattern to
  reduce the amount of bookkeeping required of its users

- converts the ingest-service to using the SingleTaskWorkflow pattern

- removes many instances of the word "job." This was likely introduced
  because pre-cereal implementations used the word "job" and in our
  first pass of converting them to cereal we aimed for minimal
  changes. "Job" is confusing since cereal doesn't have a concept
  called "job" and thus what it refers to is unclear. The word "job"
  still appears in much code as it is part of various APIs.

**Implementations Pros and Cons**

For this we, considered two implementations:

*Implementation 1*

```golang
type WorkflowName string
type TaskTame string
```

The advantages of this approach were:

- Easy to construct: `cereal.WorkflowName("foo")`

- TaskName's and WorkflowName's can be constants

- Directly comparable to strings

The disadvantages of this approach were:

- String literals ("foo") and untyped string constants (`const foo =
  "foo"`) implicitly convert to these types, meaning many of the
  mistakes we want to avoid are still possible with no compiler
  warning.

*Implementation 2* (we did this one)

```golang
type WorkflowName struct { value string}
type TaskName struct { value string}

func NewWorkflowName(s string) TaskName { return TaskName{value: s} }
func NewTaskName(s string) TaskName { return TaskName{value: s} }
```

The advantages of this approach were:

- No implicit conversion from string literals or untyped constants to
  this type are allowed, meaning that the compiler catches all of the
  cases we care about.

The disadvantages of this approach were:

- Requires a constructor to avoid other lint warnings

- Can't be assigned to a constant, thus constants need to be converted
  to package-level vars

- Direct comparison with strings doesn't work

We originally thought that some of the ergonomic advantages of
Implementation 1 might be worth it and that we could develop a static
analyzer to find the problematic implicit conversion cases. Reviewers
can see this implementation in the first few commits. After looking at
the result, I decided that the small ergonomic problems with the
struct-based implementation 2 weren't bad enough to merit maintaining
the linter which would have had to handle a number of edge cases.

### Future Work

- Reviewers will note that we almost immediately convert
  `WorkflowName` and `TaskName` back into strings inside the
  library. To fix this, we would have to unify some of the types used
  between the `backend` package and the `cereal` package. Since this
  may also require substantial changes to the package layout, we have
  pushed this to a potential follow-up PR.

- Another pass to remove "job" from any place where it isn't
  semantically correct might be required.

- "Instance Name" appears to be used in multiple contexts. We should
  review this and split it into two different names if appropriate.

- "Instance Name" is still a string in the public APIs. As this is the
  only string, I'm less worried about it, but we can do a similar
  conversion to a specific type if required.

- The `Schedule` type is shared between the backend and frontend and
  as such still uses `string`.

Signed-off-by: Steven Danna <steve@chef.io>