render-template: add ability to read from the "secrets-helper"
The "secrets-helper" is a poorly named utility for sharing data
between services.  The data inside stored in it is not stored
security, but it is named "secrets-helper" because it conforms to an
older secrets-related API in the Chef Server.

In templates, you can now do:

    {{ getSecret "foo.bar" }}

to get data out of this store.

Signed-off-by: Steven Danna <steve@chef.io>