[AUTH-1471] Project deletion from domains via cereal (#1670)
Make project deletion more transactional across domains using cereal. Now projects will also be purged from non-authz domains on project deletion. Now the process is:

Serial: Move the project to a graveyard db table. Will 404 or 500 right away if the project doesn't exist or there is a problem. This part is idempotent if the project has already been moved to the graveyard to handle the case where the move happens successfully and the db transaction is committed but we crash before we report the success to cereal (the db function will be retried and return no error).

Parallel: Delete the project from every other non-authz domain (currently teams and tokens, but wired in a way that we can easily add more in the future). On failure, it will retry with exponential backoff and log. We don't expect long term failures here unless the system is unhealthy.

Serial: After the project has been deleted from every domain, we remove it from the graveyard, at which point the ID can be re-used for a new project since it has been completely purged from the system.