Jamie/fix datetime compliance reports (#1821)
* Add ChefDateTime string for use across application

Currently named ChefDateTime (a WIP) to use across the app as a
time format that will be regularly reused.  I also needed to
import the custom chef pipes to be able to use in this module.

Signed-off-by: seajamied <jdegnan@chef.io>

* Update D3 time format to match requested ChefDateTime

Signed-off-by: seajamied <jdegnan@chef.io>

* Add RFC2822 time to reporting-node component

This required adding the ChefPipesModule to the component module as well.  I also removed
a no longer needed formateDate function from the reporting-node component.

Signed-off-by: seajamied <jdegnan@chef.io>

* Import DatetimePipe to reporting.component spec

Signed-off-by: seajamied <jdegnan@chef.io>

* Import DatetimePipe into reporting-node spec

Signed-off-by: seajamied <jdegnan@chef.io>

* Favor moment over d3 time formatting in overview-trend

Signed-off-by: seajamied <jdegnan@chef.io>