Major overhaul on showInputs()
Completely reworked so that form controls are being removed or added to the FormGroup, instead of
shown or hidden.  This allows the Validations to stay consistent between adding/removal instead of
flipping them of and off.  This also allows the ngIf to display based on whether or not a formControl
exists, which is the suggested approach by the angular core team https://github.com/angular/angular/issues/7970\#issuecomment-228624899

Signed-off-by: seajamied <jdegnan@chef.io>