deployment: use config information for dynamic port preflight checks
The preflight checks were previously a hard-coded list of ports. This
list was out of date, did not include ports from optional collections,
and only tested the default port even if the user had changed their
configuration.

Now, we use the AutomateConfig struct and related interfaces to check
all ports, using any custom user configuration.

As a result of this change, in the default configuration, the
following unused ports will no longer be checked:

2133
2134
2192
10114
10118
10159

And the following new ports will be checked:

 4222 -> event-gateway
10101 -> cereal
10134 -> event-feed-service
10140 -> event-service
10130 -> authz-service
10132 -> event-service
10143 -> backup-gateway
10144 -> es-gateway
10145 -> pg-gateway
10147 -> event-service
10148 -> event-gateway
14001 -> data-feed-service

Further, when the user configures an optional product, the ports for
those products will also be checked.

Signed-off-by: Steven Danna <steve@chef.io>