preflight: skip workflow ports in check during a1migration
These ports were previously unchecked. They were automatically
included in the preflight checks when we moved to dynamic port
checking; however, they need to be skipped during the a1 migration
because they are also used by A1.

Note, that while we could move port 9611, I'm not sure we can move 8989
as this is also used in the people's git-remote configuration.

Further note that automate-workflow-nginx's ports are currently
hard-coded and not included as part of the dynamic check.

Signed-off-by: Steven Danna <steve@chef.io>