ci: Add test for deploying with the unstable Habitat build
We would like to be able to upgrade to the latest version to hab much
more quickly than we currently do. One issue that causes delay is that
we often don't test the latest version of Habitat until after it is
released. If we find an issue that we need to address upstream in
Habitat, this delays our upgrade as we need to submit the fix, wait
for the release, and then re-test.

The goal of this change is to get us continuously testing the dev
builds in unstable so we can catch issues or confirm fixes before the
release of Hab.

Signed-off-by: Steven Danna <steve@chef.io>