Fix bug where duplicate projects getting returned from OPA in 2.1 could result in * (#1963)
This is an edge case where the duplicates would be the same length as the total projects in the system.

For example, if you have a single custom project named project1 one and two policies that allow user bob on project1 for a specific resource and action, then V2ProjectsAuthorized would return ["project1", "project1"]. Since the total projects in the system would be ["(unassigned)", "project1"], ProjectsAuthorized would incorrectly equate the result of V2ProjectsAuthorized to * due to the fact that their lengths matched. This edge case would happen anytime NumberOfProjectsAllowingAProjectForASpecificResourceAndAction = NumberOfCustomProjects + 1 (for the unassigned project).

V2ProjectsAuthorized has been updated to not return duplicates which was never the intended behavior. The duplicates were introduced in this change:

https://github.com/chef/automate/commit/c21ddbe4cfd751f0b5513211a439caaef6cddabf

Due to the introduction of using a partial document:

https://github.com/chef/automate/commit/c21ddbe4cfd751f0b5513211a439caaef6cddabf#diff-2d317bb33ff7babca7e1290b0fb1dcf2R241

Which unfortunately contain duplicates:

https://github.com/open-policy-agent/opa/issues/429

Signed-off-by: Tyler Cloke <tylercloke@gmail.com>