Project update for compliance and client runs elasticsearch bug (#2024)
When running a project update with a chef tag project rule the compliance-service would fail to update. The problem is that the array field is sometimes stored as "null" and not an empty array. The solution is to check if the field is null before using it.

Client runs had the same problem which is fixed in the PR also.

Signed-off-by: Lance Finfrock <lfinfrock@chef.io>