Build empty data lifecycle service
We create an empty build because deployment-service cannot continue if we
remove a service from the manifest. The empty package is the least  bad of
our options because it has no dependencies, so it wont cause us to install
in data-lifecycle-service's old deps.

This will be a 2 part PR. We'll pin data-lifecycle-service in
products.meta once its built.

Signed-off-by: Jay Mundrawala <jmundrawala@chef.io>