Allow proxy setup in integration test framework
This change allows us to test proxy setups directly
in the integration test framework.

A test can declare that it wants to test proxies with:

```bash
test_proxy=true
```

This will setup a container running squid. The container running
automate will have access to the squid container, but not the internet.
The squid proxy will have access to the internet.

do_setup exports the HTTP(S)_PROXY environment variables.

Signed-off-by: Jay Mundrawala <jmundrawala@chef.io>