Removing Duplicate requests for details pages (#2113)
Updating the policy details and team details pages to not request data when the page's internal tabs are changed. This solution combines two observers that are triggered off URL updates. Added the "distinctUntilChanged" filter to prevent extra data request when the internal tabs are changed. The "distinctUntilChanged" filter prevents multiple events from being sent when the data has not changed.

Signed-off-by: Lance Finfrock <lfinfrock@chef.io>