[AUTOMATE-2138] Always refetch latest data in UI for projects, policies, and roles
When calls to `GET` and `GET_ALL` happen in the store, clear the store of the object in question and refetch so that we always have the latest data.

Also updated all the subscriptions to filter until the fetch is completed to avoid flickering and other undesirable side effects of hitting the subscribe code multiple times.

Signed-off-by: Tyler Cloke <tylercloke@gmail.com>