gateway: quick fix for mockgen Make targets
This fixes a chicken-and-egg problem that we hit when trying to add a
service since the package the code-generator reflects on also requires
that the generated code is already correctly generated.

Note the change in clients_mock.go is likely from the fact that this
file was hand-edited in the past.

I've also removed a Make target which simply didn't work because the
relevant files didn't exist.

Further, I noticed that running `make` produces files in

    mock_compliance_events/

which are not currently in source control. I haven't added them here
as as I wasn't sure if this was intentional.

In short, mock generation is a big mess still and needs to be cleaned
up, but at least now `make` works again.

Signed-off-by: Steven Danna <steve@chef.io>
Signed-off-by: Vivek Singh <vivek.singh@msystechnologies.com>