ci: update automation of habitat version bumping
The Habitat release process has changed. They now publish a manifest
with all relevant package versions.

By default the script will pull from the latest stable release, but
TARGET_VERSION can be set to an exact version for local testing or
manually setting the pin.

The workload I've subscribed to was taken from:

  https://github.com/habitat-sh/homebrew-habitat/pull/28

The script has been testing locally, but unfortunately we don't have a
great way to confirm that it will work in the context of expeditor.

Signed-off-by: Steven Danna <steve@chef.io>