[data-lifecycle] Migrate default values from old config
The data lifecycle retention migration logic used in compliance,
ingest, and event-feed was written before the default configuration
removal. Because of the changes in the default values after the default
configuration was removed, the migration logic was invalid.

With this change we modify the default values that are used when running
the aforementioned services to be our desired defaults. We also handle
cases where a user had previously configured services with their own
custom value.

Signed-off-by: Ryan Cragun <ryan@chef.io>