Embed latest nats-server in event-gw and event-svc
We had some patches to nats-server to support mutual TLS, which have
since been merged. In the meantime, the nats project shipped version 2
(which was very soon after our initial import/vendor) and we moved
automate from the `dep` tool to modules. This patch contains the
necessary changes to use the 2.x nats-server code in the event-gateway
and event-service. End-to-end functionality was verified using the
docker-based hab-client from applications-service's dev/ dir.

Signed-off-by: Daniel DeLeo <dan@chef.io>