studio: don't bind mount /
We believe this was put in place to get around various checks in ES
that explicitly look for / in the mount table.

However, this was inserted a while ago, recently made some debugging
more complicated, and may no longer be required.

If CI passes, I think we should remove it, if it becomes an issue in
everyday studio use, we can bring it back.

Signed-off-by: Steven Danna <steve@chef.io>