nodemanager Read RPC project filtering (#2347)
Adding project filtering to the "api/v0/nodes/"/Read RPC function. This is needed because the nodemanager stores both ingested nodes and manually added nodes. These changes filter the ingested nodes only. The manually added nodes will not be filtered by the project tags in the context object.

If the ingested node is not allowed by the passed in projects in the request a 404 will be returned.

Signed-off-by: Lance Finfrock <lfinfrock@chef.io>