Remove all applied search query filters when user clicks `clear all` button (#2306)
* Send an empty object when removing all filters

Signed-off-by: seajamied <jdegnan@chef.io>

* Return false early if key is one of control_tag

This works and is somewhat pragmatic, though i'm not sure this is the final version yet.
The current format works by looping through all of our allowed filters and then removing them
if they exist.  However, when we added 3rd level filtering for control tags, we include a : for
delimiting.  This now would either require us to manually add in each of those options to
the allow filtering array, or come up with a more dynamic version.

Signed-off-by: seajamied <jdegnan@chef.io>