Fail purge if not executed within 10 mins of being enqueued
There was a bug in a release of automate where purge workflows
would be queued up, but would not run because nothing was listening for
them. When we fixed this, people were confused by why automate was
deleting data. More confusingly, the policy that would be used was the
old one and not the current one.

Signed-off-by: Jay Mundrawala <jmundrawala@chef.io>