[AUTOMATE-2260] If the user has pending changes in the projects filter, do not use polled updates
We have two minute polling to get updated projects populated in the projects filter dropdown. However, if that polling triggers while edits are pending in the dropdown, it would remove the pending changes. Only update the dropdown if there are not pending changes so that user's edits aren't blown away by polling updates.

Signed-off-by: Tyler Cloke <tylercloke@gmail.com>