Link to builder for applications tab
This change adds a new configuration lifecycle method: ConfigurProducts
where a service can do things with the products that are deployed

load balancer uses this to render a script with the deployed products

The UI reads this script, and uses that to figure out what to render for
the link. If builder is deployed, we link to the internal builder.
Otherwise, we link to bldr.habitat.sh

Signed-off-by: Jay Mundrawala <jmundrawala@chef.io>