fix revendor target and ignore EINVAL in atomic writer
This fixes two problems that led to `revendor` not working correctly.

Previously, we were copying files from grpc-ecosystem/grpc-gateway
manually into our vendor directory. However, to determine the path to
copy from, we needed to look up the the path after installation. But,
because of some confusion about how `$(eval)` works, this did not work
on systems where that go module was not already installed.

To fix this, I've moved this logic to a shell script since I think
more people will be able to reason about the behavior of the shell
script.

Additionally, we now ignore EINVAL from directory syncs in the atomic
writer. Vboxfs and potentially other filesystems don't support calling
sync on a directory, returning EINVAL. This has been a recurring
source of confusion. Other bits of code turn of sync'ing completely to
avoid this issue, but that is a bit overkill.

Signed-off-by: Steven Danna <steve@chef.io>