add backupRunning state to converger
The new backupRunning state represents the fact that we do not want to
run most operations while the backup is running. The converger enters
this state when then backup code runs StartBackup.

Right now, the backup code remains outside of the compiler, so we've
also added a BackupComplete() call to signal when we should leave this
state.

Attempts to StartBackup() when we are waiting for a restart or a
reconfigure will fail. The client then retries these failures.

Signed-off-by: Steven Danna <steve@chef.io>