Bug fix: reset error indicator for subsequent use
After you attempt to create a rule whose id already exists,
a popup on the create-rule page indicates this error.
If you then cancel to close the page, then again select "create rule",
that error popup was still present.
This fixes that and streamlines a couple other things as well.

Signed-off-by: michael sorens <msorens@chef.io>