Remove services view state from URL and update service groups URL (#2503)
Updates the URLs in the apps tab as specified in #1884

Removes services view state from the URL; instead it is managed in the application memory. This makes sense architecturally because a particular service group can come and go from a set of service group filters over time. For example, if the critical status filter is selected, and the services within a service group all change status to OK, then the URL /applications/service-groups?status=critical&sgId=<the id> no longer refers to the same content in the same way.

This change means that when users share URLs they won't have the same service group selection state, though if a user wishes to construct a URL that shows just a particular service group, they can still do so with the filtering functionality we have.

Making this change means that changes to service group view state will no longer trigger events from observables that watch the query parameters, which solves a bug where the service group list was being re-rendered for changes to the services view state and would lose the user's scroll position after flickering the content (#2149)