update binlinks when upgrading go in expeditor environment (#2632)
n 3cc632ce6fab8aa346de7b83e0177b27811c804a we modified the script that builds the CLI docs so it upgrades go to 1.13 in the expeditor environment. However, we didn't use `--force` to overwrite existing binlinks. We are now seeing builds fail where `.expeditor/generate-automate-cli-docs.sh` is stuck with a too-old version of go, and in the logs for those builds, we see 

```
Â» Binlinking go from core/go/1.13.5/20191212163200 into /bin
Â» Binlinking gofmt from core/go/1.13.5/20191212163200 into /bin
Ã˜ Skipping binlink because go already exists at /bin/go. Use --force to overwrite
Ã˜ Skipping binlink because gofmt already exists at /bin/gofmt. Use --force to overwrite
```

Here we add `--force` so that we get the desired effect of upgrading go.


Signed-off-by: Daniel DeLeo <dan@chef.io>