Fix delete dispatching twice
The problem here with "onSelectionChange"  is that when you select a new option, it both selects the new option and then deselects the old one. So you get one event with the new option and a second event with the old option.
The solution is to check for the new option selection only,
the one with `isUserInput` being true.
Reference: https://github.com/angular/components/issues/4094

Signed-off-by: michael sorens <msorens@chef.io>