Fixed edge case of dirty/pristine checking inputs
I discovered that if one loads a URL fresh with a page containing
an input element regulated by FormControlDirective,
the initial "original value" will be set to null.
This tweak allows capturing the slightly delayed "original value".

But why did this seem to work before?
It worked fine if you were navigating within the app,
which would be occurring most of the time.
The fix here is only needed when refreshing the browser page.

Signed-off-by: michael sorens <msorens@chef.io>