Remove state-clearing upon initiating "get all"
We may have been a bit over-eager in state clearing in a couple
earlier PRs, notably https://github.com/chef/automate/pull/2158.
I think it still makes sense for individual "get" calls,
but "get all" does not need it; when the result is returned and
added to the state with an `addAll` call, that completely replaces
the section of state at hand.

Reference: https://ngrx.io/guide/entity/adapter#adapter-collection-methods

Signed-off-by: michael sorens <msorens@chef.io>