studio: use vendor-state specific directory for GOBIN
The binaries we build and put in /src/bin are typically dependent on
the state of our vendored modules. This sets up GOBIN to be based on
the hash of `vendor/modules.txt` to help ensure we don't see errors
caused by old binaries.

We don't clean up the old bin directories by default, but we do warn
devs as they enter the studio.

Signed-off-by: Steven Danna <steve@chef.io>