Dan/bump pq (#2858)
Upgrade lib/pq. This had been blocked because this upgrade causes new test failures we did not have before because it makes pq return an EPIPE error where it had been able to transparently reconnect before. This PR does not attempt to fix the transparent reconnect in pq because another contributor had tried that and got bogged down in code review. We also don't want to be stuck maintaining a patched pq indefinitely. Therefore this patch increases the pg-gateway idle timeout to 12h by default and makes it configurable, and makes the tests more tolerant of the errors that can arise from timed-out dead connections; in the case of the ha-cluster, Automate is not in control of the idle timeout so we see more problems there.

Commits:

* bump lib/pq to 1.2.0

Full set of changes:

https://github.com/lib/pq/compare/83612a56d3dd...3427c32cb71afc948325f299f040e53c1dd78979

Highlights include:

- A fix for a potential deadlock when related to commit and rollback
- QuoteLiteral function that can replace our custom implementation

* Retry admin token creation in diagnostics
* workaround inspec tests not handling 500s from stale db conns
* Increase idle timeout for pg and make it configurable-ish
* Allow pg-gateway idle connection time to be set in automate config
* Make cfgmgmt_actions diagnostics less picky
* More cfgmgmt_actions diagnostics fixes
* Wrap of a nil error is just nil :|
* Retry 403 in diagnostics in case of auth internal error
* Dial back added debug/logging
* try a different means of purging stale connections

Co-authored-by: Steven Danna <steve@chef.io>