add sql.Open wrapper that sets ConnectionMaxLifetime and pinging
The db.PGOpen and db.PGOpenContext() helpers create a database
connection with a default maximum connection lifetime to something
less than automates pg gateway timeout. The goal of this change is to
avoid errors caused by attempts to use connections that have already
been closed by pg-gateway (which will close connections after 12 hours
of inactivity) or other middleboxes.

Since ConnectionMaxLifetime limits the total lifetime of the
connection, we likely want to avoid setting it too low since it will
force us to re-create db connections even on a busy server.

The periodic db pinger also serve to keep the idle connections
alive. Because we don't control the pooling though, it is likely we'll
still see the occasional timed-out connection.

Co-authored-by: Steven Danna <steve@chef.io>
Signed-off-by: Jay Mundrawala <jmundrawala@chef.io>
Signed-off-by: Steven Danna <steve@chef.io>