[backup] Add basic artifact repository integrity checking
The shared aritifact snapshot repository is used for builder hartifact
backup operations. The current implementation determines which artifacts
_should_ exist in the shared repository by aggregating all existing
snapshots together and doing a diff on the snapshots required and the
existing assumed snapshots. The speed of this check is desireable as it
doesn't require checking the entire repository for each backup, however,
it also has no way of knowing if the artifacts are actually still in the
shared repository.

This change introduces basic shares repository integrity validation
checking. It scans the backup hartifact repository and validates any
and/or all snapshots with the actual artifacts and saves the latest
validation check results in the snapshot bucket.

`chef-automate backup integrity validate` will validate all snapshots in
the repository. A user may also choose to exclusively validate one or
more snapshots by passing backup ID's as variable arguments to the
command.

`chef-automate backup integrity show` will pretty print the current
snapshot validation metadata. Eventually it might be nice to allow this
command to take backup ID and to show all the details of a snapshot,
including which hartifacts are missing if it is corrupted.

In the future we should look to add additional commands like `repair`
which will validate the snapshot and attempt to upload any missing
artifacts from the existing builder bucket.

This change also includes a new builder scaffolding tool called
`generate real` which is a one stop shop for priming a a development
builder instance with a package.

`go run ./tools/builder-scaffolding/... generate real ./tools/builder-scaffolding/habitat/ `
will automatically build the shim package `builder-scaffolding` and
upload it to builder. It takes care of hab key, tokens, login, origins etc.

To facilate a few functions in the new generation tool
`fileutils.CopyFile()` and `fileutils.CopyDir()` have been added.

The `replayableStream` stream type has been added which allows the same
stream to be repeatably read over and over. It's syntantic sugar over
existing streams so we can cache long I/O stream reads and replay it
locally many times. It was used to handle reading the remote bucket
objects once and replaying it for each snapshot that is validated.

Signed-off-by: Ryan Cragun <ryan@chef.io>