Full refactor of the front end and patching the forms to be enabled or disabled
I am now choosing to have each line be its own separate form.  The reason for this is that
The API does not send back 1 to 1 mapped data, there is a lot of nested data that does
not make it easy to place in different buckets - will ultimately be separating by
category AND name.

Signed-off-by: seajamied <jdegnan@chef.io>