Cleaned up docker locally for breaking change in postgres image (#3047)
Recently, the docker image's default security config was changed because it defaulted to super insecure settings:

https://github.com/docker-library/postgres/issues/580

We didn't care about that since we were only using it locally for testing. Updating the Makefiles for auth related services to work with newer versions of the docker postgres image. Also cleaned up some of authn's use of PG_URL so there was less inconsistency and hard-codedness.

Signed-off-by: Tyler Cloke <tylercloke@gmail.com>