Remove angular lifecycle hooks
Lifecycle hooks are only usable on Components and Directives.
Reference: https://angular.io/guide/lifecycle-hooks

The code in the ngOnInit hook was never being called,
which is the root cause of the bug.
The noOnDestroy hook code was never used either, for that same reason,
plus that there was no `takeUntil` using it!

Signed-off-by: michael sorens <msorens@chef.io>