Quell unit test warning
This was not causing tests to fail, but when run from the command line
(with e.g. `make unit`) this warning appeared hundreds of times:

WARN: '@ngrx/store: The feature name "policies" does not exist in the state, therefore createFeatureSelector cannot access it.  Be sure it is imported in a loaded module using StoreModule.forRoot('policies', ...) or StoreModule.forFeature('policies', ...).  If the default state is intended to be undefined, as is the case with router state, this development-only warning message can be ignored.'

Signed-off-by: michael sorens <msorens@chef.io>