[automate-3031] Show pending edits bar promptly upon login (#3100)
* Remove angular lifecycle hooks

Lifecycle hooks are only usable on Components and Directives.
Reference: https://angular.io/guide/lifecycle-hooks

The code in the ngOnInit hook was never being called,
which is the root cause of the bug.
The noOnDestroy hook code was never used either, for that same reason,
plus that there was no `takeUntil` using it!

Signed-off-by: michael sorens <msorens@chef.io>

* Fix unit test error

Not a "full" test failure
(does not show up in the browser or with wallaby)
but occurs on the console:

HeadlessChrome 80.0.3987 (Mac OS X 10.14.6) ERROR
  Uncaught TypeError: Cannot read property 'iamMajorVersion' of undefined
  at http://localhost:9876/_karma_webpack_/polyfills.js:6162:25

  TypeError: Cannot read property 'iamMajorVersion' of undefined
      at http://localhost:9876/_karma_webpack_/main.js:55636:16
      at http://localhost:9876/_karma_webpack_/vendor.js:171706:30
      at memoized (http://localhost:9876/_karma_webpack_/vendor.js:171629:40)
      at defaultStateFn (http://localhost:9876/_karma_webpack_/vendor.js:171661:43)
      at http://localhost:9876/_karma_webpack_/vendor.js:171715:36
      at memoized (http://localhost:9876/_karma_webpack_/vendor.js:171629:40)
      at MapSubscriber.project (http://localhost:9876/_karma_webpack_/vendor.js:171540:23)
      at MapSubscriber._next (http://localhost:9876/_karma_webpack_/vendor.js:359692:35)
      at MapSubscriber.next (http://localhost:9876/_karma_webpack_/vendor.js:355197:18)
      at State._subscribe (http://localhost:9876/_karma_webpack_/vendor.js:354448:24)
      at ____________________Elapsed_11_ms__At__Sat_Mar_14_2020_23_08_42_GMT_0700__Pacific_Daylight_Time_ (http://localhost)
      at Object.onScheduleTask (http://localhost:9876/_karma_webpack_/vendor.js:367036:22)
      at ZoneDelegate.scheduleTask (http://localhost:9876/_karma_webpack_/polyfills.js:6362:51)
      at Object.onScheduleTask (http://localhost:9876/_karma_webpack_/polyfills.js:6262:65)
      at ZoneDelegate.scheduleTask (http://localhost:9876/_karma_webpack_/polyfills.js:6362:51)
      at Zone.scheduleTask (http://localhost:9876/_karma_webpack_/polyfills.js:6201:43)
      at Zone.scheduleMacroTask (http://localhost:9876/_karma_webpack_/polyfills.js:6224:25)
      at scheduleMacroTaskWithCurrentZone (http://localhost:9876/_karma_webpack_/polyfills.js:7097:25)
      at http://localhost:9876/_karma_webpack_/polyfills.js:8676:30

Signed-off-by: michael sorens <msorens@chef.io>

* Quell unit test warning

This was not causing tests to fail, but when run from the command line
(with e.g. `make unit`) this warning appeared hundreds of times:

WARN: '@ngrx/store: The feature name "policies" does not exist in the state, therefore createFeatureSelector cannot access it.  Be sure it is imported in a loaded module using StoreModule.forRoot('policies', ...) or StoreModule.forFeature('policies', ...).  If the default state is intended to be undefined, as is the case with router state, this development-only warning message can be ignored.'

Signed-off-by: michael sorens <msorens@chef.io>