Make pending edits apply to multi-user environment
The prior fix (https://github.com/chef/automate/pull/3100)
made "pending edits" status appear promptly for a newly logged in user,
but neglected to handle if edits become pending while a user is logged in.

This modifies the rules status polling to also include a project fetch,
which is used to determine whether edits are pending.
Also changing the interval from 1 to 2 minutes just to keep
overhead to a minimum.

Signed-off-by: michael sorens <msorens@chef.io>